enhance(upgrader): gracefully replace the old exe with the new exe
