+++ app/version.json
@@ -1,1 +1,1 @@
-{"version":"2.0.0-beta.26","build_id":142,"total_build":346}
+{"version":"2.0.0-beta.26","build_id":144,"total_build":348}
